+++ NEWS.md
@@ -19,6 +19,17 @@
 
 * `deployApp()` with `logLevel = "verbose"` no longer errors using the `httr2` backend. (#1312)
 
+* `deployApp()`, `writeManifest()`, and `appDependencies()` gain a
+  `dependencyResolution` parameter. Set `dependencyResolution = "library"` to ignore
+  the `renv.lock` file and resolve package dependencies by scanning the code.
+  The version that is recorded is what is installed in the libraries active in the
+  R session (i.e. what is displayed with `.libPaths()`). This is useful when
+  deploying from environments where there is a mismatch between the
+  renv lock file and the user's environment. (#1046, #1315, #1317)
+
+* Improved error messages when `renv::snapshot()` fails during dependency
+  discovery. (#1078)
+
 # rsconnect 1.8.0
 
 * `rsconnect` now uses [`httr2`](https://httr2.r-lib.org/) as its HTTP client.